Feat/integration

+++ server/routes/teams.js
@@ -107,7 +107,9 @@ router.get(
     });
     console.log("team--->", team);
     if (!team) {
-      res.status(response.noTeamExists.statusCode).send({ error: response.noTeamExists.message });
+      res
+        .status(response.noTeamExists.statusCode)
+        .send({ error: response.noTeamExists.message });
     }
     res.json(team);
   })
@@ -157,7 +159,9 @@ router.post(
       },
     });
     if (userteam) {
-      res.status(response.userAlreadyExists.statusCode).send({ error: response.userAlreadyExists.message });
+      res
+        .status(response.userAlreadyExists.statusCode)
+        .send({ error: response.userAlreadyExists.message });
     } else if (!userteam) {
       const newUserTeam = await UserTeam.create({
         team_id: team_id,
@@ -167,7 +171,6 @@ router.post(
     }
   })
 );
-
 
 // Delete a user from a team
 router.delete(
@@ -187,21 +190,26 @@ router.delete(
 
       if (!userteam) {
         // If the user is not a member, return a "Not Found" response
-        return res.status(response.notFound.statusCode).json({ error: 'User not found in the team' });
+        return res
+          .status(response.notFound.statusCode)
+          .json({ error: "User not found in the team" });
       }
 
       // If the user is a member, delete the UserTeam association
       await userteam.destroy();
 
       // Send a success response
-      res.status(response.ok.statusCode).json({ message: 'User removed from the team' });
+      res
+        .status(response.ok.statusCode)
+        .json({ message: "User removed from the team" });
     } catch (error) {
       console.error(error);
-      res.status(response.internalServerError.statusCode).json({ error: response.internalServerError.message });
+      res
+        .status(response.internalServerError.statusCode)
+        .json({ error: response.internalServerError.message });
     }
   })
 );
-
 
 //Edit team description
 router.put(
@@ -219,6 +227,25 @@ router.put(
         },
       }
     );
+  })
+);
+//Edit team name
+router.put(
+  "/:teamId/name",
+  asyncHandler(async (req, res, next) => {
+    const team_id = req.params.teamId;
+    const { name } = req.body;
+    await Team.update(
+      {
+        name: name,
+      },
+      {
+        where: {
+          id: team_id,
+        },
+      }
+    );
+    res.status(response.ok.statusCode).json({ message: "team name updated" });
   })
 );
 
@@ -248,14 +275,46 @@ router.post(
 
 //Delete team
 router.delete(
-  "/:teamId/",
+  "/:teamId/userTeam/:userId",
   asyncHandler(async (req, res, next) => {
     const team_id = req.params.teamId;
-    const project_id = req.params.projectId;
-    const team = await Team.destroy({
-      where: { id: team_id },
+    const user_id = req.params.userId;
+    console.log("user_id--->", user_id);
+    console.log("team_id--->", team_id);
+    // Check if there are any projects associated with the team
+    const projectsCount = await Project.count({
+      where: { team_id: team_id },
     });
-    res.status(response.ok.statusCode);
+
+    // Check if there are any users associated with the team
+    const usersCount = await UserTeam.count({
+      where: { team_id: team_id },
+    });
+
+    if (usersCount === 1 && projectsCount < 1) {
+      // If there's only one user, remove the foreign key reference
+      await UserTeam.destroy({
+        where: { team_id: team_id, user_id: user_id },
+      });
+
+      // Then, proceed with deleting the team
+      await Team.destroy({
+        where: { id: team_id },
+      });
+
+      return res.status(204).end(); // 204 No Content response for successful deletion
+    } else if (usersCount > 1) {
+      // If there are multiple users, send an error response
+      return res.status(400).json({
+        message: "Cannot delete team with multiple users, first remove users!",
+      });
+    } else if (projectsCount > 0) {
+      // If there are projects, send an error response
+      return res.status(400).json({
+        message:
+          "Cannot delete team with associated projects,delete the projects first to delete the team.",
+      });
+    }
   })
 );
 
@@ -266,7 +325,9 @@ router.post(
       const { email, teamId, invitedBy } = req.body;
 
       if (!email || !teamId || !invitedBy) {
-        return res.status(response.fieldsMissing.statusCode).json({ message: response.fieldsMissing.message });
+        return res
+          .status(response.fieldsMissing.statusCode)
+          .json({ message: response.fieldsMissing.message });
       }
 
       const [userExistInvitation, user, team] = await Promise.all([
@@ -287,22 +348,27 @@ router.post(
         return res.status(409).json({ message: errorMessage });
       }
 
-      const invited = await Invitations.create({ email, team_id: teamId, invited_by: invitedBy });
+      const invited = await Invitations.create({
+        email,
+        team_id: teamId,
+        invited_by: invitedBy,
+      });
       sendEmail({
-        "to": email,
-        "subject": "Invite for join WorkPlace",
-        "templateName": "inviteUserWorkplace", // Name of the EJS template file without the ".ejs" extension
-        "templateData": {
-          "workplaceName": team.Projects===[]?team.Projects[0].name:null,
-          "inviterName": user.name
-        }
-      })
-    
+        to: email,
+        subject: "Invite for join WorkPlace",
+        templateName: "inviteUserWorkplace", // Name of the EJS template file without the ".ejs" extension
+        templateData: {
+          workplaceName: team.Projects === [] ? team.Projects[0].name : null,
+          inviterName: user.name,
+        },
+      });
 
       res.json(invited);
     } catch (error) {
       console.error(error);
-      res.status(response.internalServerError.statusCode).json({ message: response.internalServerError.message });
+      res
+        .status(response.internalServerError.statusCode)
+        .json({ message: response.internalServerError.message });
     }
   })
 );
